Some improvements
